+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+- Allow early init of instrumentations (experimental).
+
 ## 1.72.1
 - Improve announce cycle by polling for the "agent ready" state once a second, this will bring the delay from "announce" to "agent ready" down from 10 to ~1 seconds for the majority of cases (in particular with the improved agent to be released soon).
 - Enforce minimum delay of one second between sending snapshot data and the first spans.